Journey.Examples.CreditCardApplication.Compute (Journey v0.10.31)

This module contains the business logic for the Credit Card Approval application, things like fetching the customer's credit score, making and communication the credit decision, etc.

Functions

all_done(values)

This function marks the flow as completed when it's all done.

choose_the_time_to_archive(values)

This function schedules archiving the execution.

choose_the_time_to_send_reminder(values)

This function simulates scheduling sending a reminder to preapproved customers.

compute_decision(map)

This function simulates computing the credit decision, based on the credit score.

fetch_credit_score(values)

This function simulates fetching a credit score from an external service.

request_credit_card_issuance(values)

This function simulates initiating the issuance and mailing of a credit card.

send_card_mailed_notification(values)

This function simulates emailing the customer and telling them that the card has been mailed.

send_congrats(values)

This function simulates sending the customer an email when their application was approved.

send_preapproval_reminder(values)

This function simulates sending the preapproved customer a reminder to request a credit card.

send_rejection(values)

This function simulates sending the customer an email when their application was declined.